Poliscope-Focus-Monitor-main/self-study-assistant/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
pillow==10.2.0
pygame==2.5.2


